requirements.txt
astroid==1.5.3
autopep8==1.3.3
certifi==2017.11.5
coverage==4.4.2
future==0.16.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
py==1.5.2
pycodestyle==2.3.1
pylint==1.7.4
pytest==3.2.5
python-telegram-bot==8.1.1
six==1.11.0
wrapt==1.10.11
